Progress update
Almost done with the run method for the Main class. Just need to run recursively and test it out. Testing will have to hold off until after fine tuning since the run method relies on the AI outputting the contents of a JSON file every time.
Will work on the data collection method. Don't have a concrete plan for it yet but it shouldn't be too hard.