Finalized main version one
Major changes:
Finished all essential logic. This is still version one so there may be some ways to optimize and make the system smarter but I've tested this with another LLM and this works with the target website. Now I need to handle cases with other websites but I believe that it won't be too much of a problem. Finding unnecessary information or web pages will make the system slow but it isn't really an issue compared to missing necessary information or web pages which definitely is.
One of the new systems I've incorporated is scraping emails, phone numbers, and dates off websites to feed into the contact and dates section respectively as context. This can be found in the scrape_html file.

Minor changes:
general-info -> overview
Removed generate_data since that's already a method in Main
Changed expected AI response
Added a few keywords and removed a few redundant keywords

Future:
Again and probably for the last time, I need to fine tune the model. While testing I can collect data and I am collecting data so the first step is already handled. The only problem is that fine tuning takes a while but we'll see how that goes.