Minor fixes + data collection
Minor fixes:
Fixed up the prompt by making sure it writes the name of the grade right after "Rising" if needed.
Specified to write dates as mm-dd-yyyy
Removed self.schema and just made it so that the context includes the incomplete response with the previous information filled out.
Fixed a bug where the entire html_contents wouldn't be sent to extract all the links and renamed the variable to raw_soup for clarity (and because it's raw soup, get it?).
Included a MAJOR keyword that left a blind spot to whether or not internships were virtual ('online')
Added FAQ to all link keywords because FAQs usually contain a LOT of information.
Oh yeah and fixed the logic for if dates was completely filled or not.

Future:
I need to cut up the entire process into even smaller chunks because the system is missing some information it could have received. In short, the system right now is completely bare bones which is to be expected as this is only version one.
I've collected data from 3 sites so far, each of which had around 7 - 10 entries so I have 23 entries of data. I will use this to create a test fine-tuned model to see if the fine tuning works and to estimate how much more data we need.