Strayed from fine-tuning
Fine tuning was taking too long, we've decided to bite the bullet and buy credits for an API instead. I've primed the codebase to expect to receive responses from an intelligent model and thus have added a "bulk mode" which asks the model to fill in all of the information for the entire schema. This will only be applied in the first recursion depth and individual information will again be collected for every recursive step afterwards.