Added output parsing capabilities
Since the model returns the JSON as a string, I added the handle_output method which will parse the string for the JSON and update self.response based on the output.

Future:
I should split this method into two, one being parse_output and the other being handle_output because this does do a lot at a time as it currently is.
Also we're buying the OpenAI API key in just a few.