Finished alpha development
I hope that's what alpha means.

Anyway the backend works to the point where if you input a valid endpoint URL, API key, and model, you can input a URL to an internship and if the model is accurate enough (I use grok-3-mini) then it will spit out an accurate completed schema.

Future:
Make sure the README instructions are multi-platform compatible, maybe even create a batch file for setup.
Add an option outside of editing the main file to choose which endpoint to target (between OpenAI, Anthropic, Deepseek, and the custom endpoint). Maybe detect which environmental variables are empty and choose the non-empty one.