Added excessively long queue filtering
Added a function that can rank the queue based on the amount of title keywords found in the link and contents. Also added a method to send a prompt to the model to filter the links.
Also fixed the Firebase Client so that a custom ID can be added.
Added a schema cleanup method.
Added an abbreviation check for states (NY -> New York)
Added the pprint.pp function to the logger.