Add content extractors
For some reason for the Ladder Internship, costs are listed on the apply page but the extractor still isn't adding the into the schema. I doubt this is due to the AI missing it and since I've already checked that the costs themselves are being extracted, I need to check truncate_contents to see if the costs aren't being added in the summary of the webpage. If that isn't the case it could be because the AI isn't picking up on it because of a rule but I doubt that. The least likely cause is that it isn't being scraped which I severely doubt.